+++ src/index.js
@@ -2,4 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from './routes/App.jsx'
 
-ReactDOM.render(<App />, document.getElementById('app'))
+
+const container = document.getElementById('app');
+
+const root = ReactDOM.createRoot(container);
+
+root.render(<App />);